Add co_aberration function and sunpos function bugfix
The sunpos function gave the wrong output of ecliptic longitude and obliquity of the ecliptic because they were
in degrees from the start and so didn't need to be converted again into degrees. Verified with values from sunpos.pro (IDL AstroLib).
Error was found as it propogated into the output of co_aberration function.

* src/sunpos.jl: Bugfix
* TODO.md: Remove 'co_aberration' from list.
* docs/src/ref.md: Add "co_aberration" entry to the manual.
* src/utils.jl: Add co_aberration entry to "utils".
* src/co_aberration.jl: Contains code of co_aberration function.
* test/util-tests.jl: Include tests for co_aberration.