Add euler function
This was interesting to translate, although it took more time to translate because the IDL example (2) I was using as reference for output checking
had mistakes in it.

Additionally, the reference from which the mean orbital values were taken in IDL was broken. So I found the closest
reference I could on the same site. The values are very close to the ones used in `helio` of IDL AstrolLib, except for that
of Mars.

A few slightly uncoventional method were used to implement `planet` type, however the output is accurate and type-stable.